Add try-catch knob TOPLINGDB_ENABLE_TRY_CATCH
previously for many try-catch, we enable them on release
and disable on debug.

sometimes it is useful to enable try-catch on debug, such
as MyTopling mtr.

so we add a knob macro TOPLINGDB_ENABLE_TRY_CATCH, it is
auto detected by NDEBUG if it is not pre-defined.

we use `if (1) ... if (0) try{}catch()` when try-catch
is disabled. this is very simple and clear.